Splitted JoyButtonSlot class into JoyButtonSlotXml and the original
