+++ src/joybutton.h
@@ -40,8 +40,6 @@
 
 class VDPad;
 class SetJoystick;
-class QXmlStreamReader;
-class QXmlStreamWriter;
 class QThread;
 
 class JoyButton : public QObject
