+++ src/joybuttonslot.h
@@ -24,9 +24,6 @@
 #include <QMetaType>
 #include <QVariant>
 
-class QXmlStreamReader;
-class QXmlStreamWriter;
-
 class JoyButtonSlot : public QObject
 {
     Q_OBJECT
@@ -77,9 +74,6 @@ public:
     QVariant getExtraData();
 
     bool isValidSlot();
-
-    virtual void readConfig(QXmlStreamReader *xml);
-    virtual void writeConfig(QXmlStreamWriter *xml);
     
 private:
     int deviceCode;
